+++ apps/admin/src/app/(admin)/templates/_components/GuideCard.tsx
