+++ apps/admin/src/app/(admin)/templates/_components/size-detail-form.tsx
